Extract audio MCP tools
Move audio synthesis/composition/effects tool registrations and the spatial audio
video helper into `server_tools_audio`. `server.py` remains the public façade and
re-exports moved handlers while importing the split module once for registration.

Constraint: Preserve MCP tool signatures, tool names, lazy imports, and direct imports from `mcp_video.server`
Rejected: Combine audio with AI/quality/effects tool moves | keep registration batches domain-focused
Confidence: high
Scope-risk: moderate
Tested: python3 -m pytest tests/test_public_surface.py tests/test_server.py::TestServerInitialization tests/test_audio_presets.py tests/test_ai_features.py::test_spatial_audio_missing_video tests/test_ai_features.py::test_spatial_audio_invalid_method tests/test_ai_features.py::test_spatial_audio_empty_positions -q --tb=short; python3 -m pytest tests/test_server.py tests/test_public_surface.py -q --tb=short; python3 -m ruff check mcp_video/server.py mcp_video/server_tools_audio.py; python3 -m ruff format mcp_video/server.py mcp_video/server_tools_audio.py; async smoke of mcp.list_tools() returning 83 tools including moved names
Not-tested: full non-slow suite after audio server tool extraction
